+++ src/components/square/square.css
@@ -1,0 +1,8 @@
+  
+  .box {
+    padding: 20px;
+    font-size: 150%;
+    background-color: #b5915f;
+    /* color: red; */
+    text-align: center;
+  }
